# core-service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PRAWOBIORCA",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Log User",
        "operationId": "log_user",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_log_user"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "503": {
            "description": "Service unavailable!"
          },
          "401": {
            "description": "Incorrect logging data!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh Tokens",
        "operationId": "refresh_tokens",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "503": {
            "description": "Service unavailable!"
          },
          "401": {
            "description": "Invalid or expired refresh token!"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Check Is User Logged",
        "operationId": "check_is_user_logged",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable!"
          },
          "401": {
            "description": "User not logged."
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout User",
        "operationId": "logout_user",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "503": {
            "description": "Service unavailable!"
          }
        }
      }
    },
...
    "/api/user/cases": {
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Get Cases List",
        "operationId": "get_cases_list",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CaseData"
                  },
                  "type": "array",
                  "title": "Response Get Cases List"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable!"
          }
        }
      },
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Add Case",
        "operationId": "add_case",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_case"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid",
                  "title": "Response Add Case"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/cases/{caseId}": {
      "delete": {
        "tags": [
          "cases"
        ],
        "summary": "Delete User Case",
        "operationId": "delete_user_case",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Caseid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "503": {
            "description": "Service unavailable!"
          },
          "404": {
            "description": "No case with that id!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/cases/{caseId}/documents": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Add Case Document",
        "operationId": "add_case_document",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Caseid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCaseDocument"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "503": {
            "description": "Service unavailable!"
          },
          "404": {
            "description": "No case with that id!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "Get Case Documents",
        "operationId": "get_case_documents",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Caseid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseDocument"
                  },
                  "title": "Response Get Case Documents"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/cases/documents/{documentId}": {
      "delete": {
        "tags": [
          "cases"
        ],
        "summary": "Delete Case Document",
        "operationId": "delete_case_document",
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Documentid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "503": {
            "description": "Service unavailable!"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_add_case": {
        "properties": {
          "caseName": {
            "type": "string",
            "title": "Casename"
          }
        },
        "type": "object",
        "required": [
          "caseName"
        ],
        "title": "Body_add_case"
      },
      "Body_log_user": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_log_user"
      },
      "CaseData": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CaseData"
      },
      "CaseDocument": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "caseId": {
            "type": "string",
            "format": "uuid",
            "title": "Caseid"
          },
          "presentationName": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Presentationname"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "id",
          "caseId",
          "presentationName",
          "content"
        ],
        "title": "CaseDocument"
      },
      "CurrentUser": {
        "properties": {
          "isAdmin": {
            "type": "boolean",
            "title": "Isadmin"
          }
        },
        "type": "object",
        "required": [
          "isAdmin"
        ],
        "title": "CurrentUser"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NewCaseDocument": {
        "properties": {
          "presentationName": {
            "type": "string",
            "minLength": 1,
            "title": "Presentationname"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "presentationName",
          "content"
        ],
        "title": "NewCaseDocument"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}